refactor: ChatRoomList 페이지를 Container/Presentational 패턴으로 리팩토링
